refactor: clean up connector before adding adminApi
