+++ packages/interledger/src/connector/index.ts
@@ -1,14 +1,11 @@
 import { randomBytes } from 'crypto'
 import compose = require('koa-compose')
-import { createClient } from 'tigerbeetle-node'
 import IORedis from 'ioredis'
 
-import { Config } from '../config'
 import {
   createApp,
   Rafiki,
   RafikiRouter,
-  createRatesService,
   createBalanceMiddleware,
   createIncomingErrorHandlerMiddleware,
   createIldcpProtocolController,
@@ -20,44 +17,29 @@ import {
   createIncomingThroughputMiddleware,
   createOutgoingReduceExpiryMiddleware,
   createOutgoingThroughputMiddleware,
-  createOutgoingValidateFulfillmentMiddleware
+  createOutgoingValidateFulfillmentMiddleware,
+  RatesService
 } from './core'
 import { Logger } from '../logger/service'
-import { AdminApi } from '../admin-api'
 import { AccountsService } from '../accounts/service'
 
-/**
- * Admin API Variables
- */
-const ADMIN_API_HOST = process.env.ADMIN_API_HOST || '127.0.0.1'
-const ADMIN_API_PORT = parseInt(process.env.ADMIN_API_PORT || '3001', 10)
-
-/**
- * Connector variables
- */
 const ILP_ADDRESS = process.env.ILP_ADDRESS || undefined
 const STREAM_SECRET = process.env.STREAM_SECRET
   ? Buffer.from(process.env.STREAM_SECRET, 'base64')
   : randomBytes(32)
-const pricesUrl = process.env.PRICES_URL // optional
-const pricesLifetime = +(process.env.PRICES_LIFETIME || 15_000)
-
-export interface ConnectorService {
-  adminApi: AdminApi
-  app: Rafiki
-}
 
 interface ServiceDependencies {
   redis: IORedis.Redis
+  logger?: typeof Logger
+  ratesService: RatesService
+  accountsService: AccountsService
 }
 
 export async function createConnectorService({
-  redis
-}: ServiceDependencies): Promise<ConnectorService> {
-  const log = Logger.child({
-    service: 'ConnectorService'
-  })
-
+  redis,
+  ratesService,
+  accountsService
+}: ServiceDependencies): Promise<Rafiki> {
   if (!ILP_ADDRESS) {
     throw new Error('ILP_ADDRESS is required')
   }
@@ -82,30 +64,8 @@ export async function createConnectorService({
     createClientController()
   ])
 
-  const ratesService = createRatesService({
-    pricesUrl,
-    pricesLifetime,
-    logger: Logger
-  })
   const middleware = compose([incoming, createBalanceMiddleware(), outgoing])
 
-  const tbClient = createClient({
-    cluster_id: Config.tigerbeetleClusterId,
-    replica_addresses: Config.tigerbeetleReplicaAddresses
-  })
-
-  const accountsService = new AccountsService(tbClient, Config, log)
-
-  const adminApi = new AdminApi(
-    { host: ADMIN_API_HOST, port: ADMIN_API_PORT },
-    {
-      auth: (): boolean => {
-        return true
-      },
-      accounts: accountsService
-      //router: router
-    }
-  )
   // TODO Add auth
   const app = createApp({
     //router: router,
@@ -128,8 +88,5 @@ export async function createConnectorService({
   // TODO Handle echo
   app.use(appRouter.routes())
 
-  return {
-    adminApi,
-    app
-  }
+  return app
 }
